LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/event-constructor-expected.txt


PASS can't bypass required members by omitting the dictionary entirely
PASS destination is required
PASS signal is required
FAIL all properties are reflected back assert_equals: expected (object) Element node <a></a> but got (undefined) undefined
FAIL defaults are as expected assert_equals: expected (object) null but got (undefined) undefined
PASS hasUAVisualTransition is default false
